ref(tests): split conversation Location coverage
