+++ policies/correctness-complexity.md
@@ -24,6 +24,9 @@ cost.
   change behavior safely, the design is not simple.
 - A review that finds "more correct but harder to understand" should require a
   simplification pass before merge unless the risk is urgent and documented.
+- Keep formatter output. Do not compress or undo formatting to pass the
+  1,000-line limit. Split an oversized file at a clear feature or ownership
+  boundary.
 - Tests should prove the rule at the highest useful edge. They should not encode
   every internal step of a complex implementation.
